[#216] LcdPrint: Use uint32_t for lcdPrintTime() function
uint16_t milliseconds counter may wrap on long operations, which could lead to wrong value on display

+++ src/core/drivers/LcdPrint.cpp
@@ -160,7 +160,7 @@ void lcdPrintValue_(uint16_t x, int8_t dig, uint16_t div, bool mili, bool minus)
 
 }
 
-void lcdPrintTime(uint16_t timeSec)
+void lcdPrintTime(uint32_t timeSec)
 {
     lcdPrintUnsigned(timeSec/60, 3, '0');
     lcdPrintChar(':');
